+++ NEWS.md
@@ -49,14 +49,15 @@
   * `setExperimentSamplesAssayFileData()`
   * `updateCellContents()`
 * Added function `getEntityAttributeMetadata()` to get attribute metadata for a
-  specified entity type. Aplicable for PFS v6 and greater.
+  specified entity type. Applicable for PFS v6 and greater.
 * Added function `getAttributesColumnHeaders()` to get the attribute column
-  headers for all or specified attributes of an entityType. Aplicable for PFS v6
+  headers for all or specified attributes of an entityType. Applicable for PFS v6
   and greater.
 * Added "semantic version" to warning, error and test messages to avoid
   confusion.
 * Enabled automated testing to ensure package compatibility with PFS 6.0.3.1 and
   PFS 5.3.11.
+* Added function `getExperimentSamplesWithSampleLots()` to get the sample lots and/or experiment samples for a particular experiment.
 
 ## Fixes